chore(event): use u16 as event type
The EventType enum is now only used in plugins and we'll get rid
of it there too.

+++ falco_event_serde/src/de/payload.rs
@@ -17,7 +17,7 @@ falco_event::derive_deftly_for_events! {
                 $(self.$fname.repr,)
             ];
 
-            let event_type_id = <falco_event::events::types::$ttype as EventPayload>::ID as u16;
+            let event_type_id = <falco_event::events::types::$ttype as EventPayload>::ID;
             let large_payload = match size_of::<<falco_event::events::types::$ttype as EventPayload>::LengthType>() {
                 2 => false,
                 4 => true,
